# website/openapi/engine.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HumanApproveRequest": {
        "description": "Optional body for human approval.",
        "properties": {
          "output": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional output override for reason-step approve-with-edit.",
            "title": "Output"
          }
        },
        "title": "HumanApproveRequest",
        "type": "object"
      },
      "HumanDecisionRequest": {
        "description": "Body for a HITL decision.",
        "properties": {
          "decision": {
            "description": "Human decision for this step.",
            "enum": [
              "APPROVE",
              "REJECT"
            ],
            "title": "Decision",
            "type": "string"
          },
          "error_details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional structured rejection reason.",
            "title": "Error Details"
          },
          "output": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional output override when approving a reason step.",
            "title": "Output"
          }
        },
        "required": [
          "decision"
        ],
        "title": "HumanDecisionRequest",
        "type": "object"
      },
      "HumanDecisionResponse": {
        "description": "Response for HITL decision enqueue (202 Accepted).",
        "properties": {
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "already_queued",
              "requeued"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "idempotency_key"
        ],
        "title": "HumanDecisionResponse",
        "type": "object"
      },
      "HumanRejectRequest": {
        "description": "Optional body for human rejection.",
        "properties": {
          "error_details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional structured rejection reason.",
            "title": "Error Details"
          }
        },
        "title": "HumanRejectRequest",
        "type": "object"
      },
      "HumanRetryRequest": {
        "description": "Optional body for manual HITL retry.",
        "properties": {
          "guidance": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Operator note for this retry run, passed to the worker as _hitl_retry.guidance (overrides the step manifest hitl_retry_guidance for this attempt).",
            "title": "Guidance"
          },
          "retry_token": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 8,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional idempotency token for this HTTP request; omit to generate a new token on each call.",
            "title": "Retry Token"
          }
        },
        "title": "HumanRetryRequest",
        "type": "object"
      },
      "HumanRetryResponse": {
        "description": "Response for HITL step retry enqueue (202 Accepted).",
        "properties": {
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          },
          "retry_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Token"
          },
          "status": {
            "enum": [
              "queued",
              "already_queued",
              "requeued"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "idempotency_key"
        ],
        "title": "HumanRetryResponse",
        "type": "object"
      },
...
      "PendingReviewStepListResponse": {
        "description": "Response for GET /v1/sagas/pending-review.",
        "properties": {
          "has_more": {
            "description": "True when len(items) == limit (more pages may exist).",
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PendingReviewStepItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Total matching rows when include_total=true was requested.",
            "title": "Total"
          }
        },
        "required": [
          "items",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "PendingReviewStepListResponse",
        "type": "object"
      },
...
      "SagaInstanceItem": {
        "description": "One row in GET /v1/sagas (instance list).",
        "properties": {
          "definition_id": {
            "title": "Definition Id",
            "type": "string"
          },
          "definition_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Definition Name"
          },
          "definition_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Definition Version"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "parent_trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Trace Id"
          },
          "start_idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Idempotency Key"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "namespace",
          "definition_id",
          "status",
          "started_at"
        ],
        "title": "SagaInstanceItem",
        "type": "object"
      },
      "SagaInstanceListResponse": {
        "description": "Response for GET /v1/sagas.",
        "properties": {
          "has_more": {
            "description": "True when len(items) == limit (more pages may exist).",
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SagaInstanceItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Total matching rows when include_total=true was requested.",
            "title": "Total"
          }
        },
        "required": [
          "items",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "SagaInstanceListResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Control-plane HTTP API for deploying manifests, starting sagas, HITL review, and operator recovery. OSS surface under /v1.",
    "title": "API Reference",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/sagas": {
      "get": {
        "description": "List saga instances, newest started_at first.",
        "operationId": "get_sagas_v1_sagas_get",
        "parameters": [
          {
            "description": "Filter by namespace; omit for all.",
            "in": "query",
            "name": "namespace",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by namespace; omit for all.",
              "title": "Namespace"
            }
          },
          {
            "description": "Filter to a single saga instance trace_id.",
            "in": "query",
            "name": "trace_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to a single saga instance trace_id.",
              "title": "Trace Id"
            }
          },
          {
            "description": "Filter to child sagas spawned by this parent trace_id.",
            "in": "query",
            "name": "parent_trace_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to child sagas spawned by this parent trace_id.",
              "title": "Parent Trace Id"
            }
          },
          {
            "description": "Repeat for multiple values, e.g. status=RUNNING&status=PENDING.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Repeat for multiple values, e.g. status=RUNNING&status=PENDING.",
              "title": "Status"
            }
          },
          {
            "description": "When true, filter to non-terminal in-flight statuses (mutually exclusive with status).",
            "in": "query",
            "name": "in_flight",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "When true, filter to non-terminal in-flight statuses (mutually exclusive with status).",
              "title": "In Flight"
            }
          },
          {
            "description": "When true, filter to FAILED sagas only (mutually exclusive with in_flight and status).",
            "in": "query",
            "name": "failed",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "When true, filter to FAILED sagas only (mutually exclusive with in_flight and status).",
              "title": "Failed"
            }
          },
          {
            "description": "Include total matching row count.",
            "in": "query",
            "name": "include_total",
            "required": false,
            "schema": {
              "default": false,
              "description": "Include total matching row count.",
              "title": "Include Total",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SagaInstanceListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Sagas",
        "tags": [
          "sagas"
        ]
      }
    },
    "/v1/sagas/pending-review": {
      "get": {
        "description": "List HITL-held steps awaiting human approval/rejection.",
        "operationId": "pending_review_steps_v1_sagas_pending_review_get",
        "parameters": [
          {
            "in": "query",
            "name": "namespace",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Namespace"
            }
          },
          {
            "in": "query",
            "name": "trace_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trace Id"
            }
          },
          {
            "description": "Optional step kind: reason or commit.",
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional step kind: reason or commit.",
              "title": "Kind"
            }
          },
          {
            "description": "Include total matching row count.",
            "in": "query",
            "name": "include_total",
            "required": false,
            "schema": {
              "default": false,
              "description": "Include total matching row count.",
              "title": "Include Total",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingReviewStepListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pending Review Steps",
        "tags": [
          "hitl"
        ]
      }
    },
...
    "/v1/sagas/{trace_id}/steps/{step_span_id}/approve": {
      "post": {
        "description": "Approve a HITL-held step and enqueue HUMAN_APPROVED for the engine.",
        "operationId": "approve_step_v1_sagas__trace_id__steps__step_span_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "trace_id",
            "required": true,
            "schema": {
              "title": "Trace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "step_span_id",
            "required": true,
            "schema": {
              "title": "Step Span Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "namespace",
            "required": false,
            "schema": {
              "default": "default",
              "title": "Namespace",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/HumanApproveRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanDecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Step not found."
          },
          "409": {
            "description": "Conflict (e.g. step not awaiting human review)."
          },
          "422": {
            "description": "Invalid request."
          }
        },
        "summary": "Approve Step",
        "tags": [
          "hitl"
        ]
      }
    },
    "/v1/sagas/{trace_id}/steps/{step_span_id}/decision": {
      "post": {
        "description": "Submit one idempotent human decision for a HITL-held step.",
        "operationId": "decide_step_v1_sagas__trace_id__steps__step_span_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "trace_id",
            "required": true,
            "schema": {
              "title": "Trace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "step_span_id",
            "required": true,
            "schema": {
              "title": "Step Span Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "namespace",
            "required": false,
            "schema": {
              "default": "default",
              "title": "Namespace",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HumanDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanDecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Step not found."
          },
          "409": {
            "description": "Conflict (e.g. step not awaiting human review)."
          },
          "422": {
            "description": "Invalid request."
          }
        },
        "summary": "Decide Step",
        "tags": [
          "hitl"
        ]
      }
    },
    "/v1/sagas/{trace_id}/steps/{step_span_id}/reject": {
      "post": {
        "description": "Reject a HITL-held step and enqueue HUMAN_REJECTED for the engine.",
        "operationId": "reject_step_v1_sagas__trace_id__steps__step_span_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "trace_id",
            "required": true,
            "schema": {
              "title": "Trace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "step_span_id",
            "required": true,
            "schema": {
              "title": "Step Span Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "namespace",
            "required": false,
            "schema": {
              "default": "default",
              "title": "Namespace",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/HumanRejectRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanDecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Step not found."
          },
          "409": {
            "description": "Conflict (e.g. step not awaiting human review)."
          },
          "422": {
            "description": "Invalid request."
          }
        },
        "summary": "Reject Step",
        "tags": [
          "hitl"
        ]
      }
    },
    "/v1/sagas/{trace_id}/steps/{step_span_id}/retry": {
      "post": {
        "description": "Re-run a HITL-held step; preserves upstream saga context (state-preserved retry).",
        "operationId": "retry_step_v1_sagas__trace_id__steps__step_span_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "trace_id",
            "required": true,
            "schema": {
              "title": "Trace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "step_span_id",
            "required": true,
            "schema": {
              "title": "Step Span Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "namespace",
            "required": false,
            "schema": {
              "default": "default",
              "title": "Namespace",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/HumanRetryRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanRetryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Step not found."
          },
          "409": {
            "description": "Conflict (e.g. step not awaiting human review)."
          },
          "422": {
            "description": "Invalid request."
          }
        },
        "summary": "Retry Step",
        "tags": [
          "hitl"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local engine",
      "url": "http://127.0.0.1:8000"
    }
  ]
}